start tdd for vec3

+++ ray_tracer.go
@@ -20,8 +20,8 @@ func main() {
 	// Set color for each pixel.
 	for x := 0; x < width; x++ {
 		for y := 0; y < height; y++ {
-			r := float64(x) / float64(width)
-			g := float64(y) / float64(height)
+			r := float32(x) / float32(width)
+			g := float32(y) / float32(height)
 			img.Set(x, y, color.RGBA{toZero255(r), toZero255(g), toZero255(0.2), 0xff})
 		}
 	}
@@ -31,6 +31,6 @@ func main() {
 	png.Encode(f, img)
 }
 
-func toZero255(x float64) uint8 {
-	return uint8(math.Round(255.99 * x))
+func toZero255(x float32) uint8 {
+	return uint8(math.Round(float64(255.99 * x)))
 }
